Reworked Animations
Animations can now be made with a general constructor in AnimationImage

Reworked a PhysicalObject's Image to be a new class called a Sprite

Sprites support animated textures! Now any physical object can have an animated texture and work fine.

Added a new weapon/weapon animation for an ion type interceptor. Try the PDI out to see it.

When reading from the image getSprite().getImage() now instead of getImage()

When writing, the method needs to be created in getSprite() to affect every image thats animated